requirements.txt
mcp
cairosvg
Pillow 